test(vitest): migrate to vitest 4 projects mode preserving app-shell behavior
Replaces vitest.config.js with vitest.config.ts using Vitest 4's test.projects
API (defineWorkspace was removed in Vitest 4). The app-shell project preserves
all existing behavior: 2908 tests, same aliases, setup files, coverage thresholds,
and pool settings. The packages/* glob now also discovers @prismgb/gpu's 19 tests.
Adds .worktrees/** exclusion to prevent sibling-branch test discovery.
