+++ scripts/swc.config.js
@@ -1,8 +1,7 @@
 /**
  * Shared SWC configuration for Vite and Vitest transpilation.
  *
- * Used by vite.config.js (including worker.plugins), vitest.config.js, and
- * (in Phase 0 Task 6) vitest.workspace.ts. Centralized to prevent drift across
+ * Used by vite.config.js (including worker.plugins) and vitest.config.ts. Centralized to prevent drift across
  * build contexts.
  *
  * Note: unplugin-swc globally disables esbuild and processes BOTH TypeScript
